Change PoS exports type architecture.
