+++ massa-pos-exports/src/cycle_info.rs
@@ -1,0 +1,49 @@
+use bitvec::vec::BitVec;
+use massa_models::{address::Address, prehash::PreHashMap};
+use num::rational::Ratio;
+use std::collections::BTreeMap;
+
+/// State of a cycle for all threads
+#[derive(Default, Debug, Clone, PartialEq, Eq)]
+pub struct CycleInfo {
+    /// cycle number
+    pub cycle: u64,
+    /// whether the cycle is complete (all slots final)
+    pub complete: bool,
+    /// number of rolls each staking address has
+    pub roll_counts: BTreeMap<Address, u64>,
+    /// random seed bits of all slots in the cycle so far
+    pub rng_seed: BitVec<u8>,
+    /// Per-address production statistics
+    pub production_stats: PreHashMap<Address, ProductionStats>,
+}
+
+/// Block production statistic
+#[derive(Default, Debug, Copy, Clone, PartialEq, Eq)]
+pub struct ProductionStats {
+    /// Number of successfully created blocks
+    pub block_success_count: u64,
+    /// Number of blocks missed
+    pub block_failure_count: u64,
+}
+
+impl ProductionStats {
+    /// Check if the production stats are above the required percentage
+    pub fn is_satisfying(&self, max_miss_ratio: &Ratio<u64>) -> bool {
+        let opportunities_count = self.block_success_count + self.block_failure_count;
+        if opportunities_count == 0 {
+            return true;
+        }
+        &Ratio::new(self.block_failure_count, opportunities_count) <= max_miss_ratio
+    }
+
+    /// Increment a production stat structure with another
+    pub fn extend(&mut self, stats: &ProductionStats) {
+        self.block_success_count = self
+            .block_success_count
+            .saturating_add(stats.block_success_count);
+        self.block_failure_count = self
+            .block_failure_count
+            .saturating_add(stats.block_failure_count);
+    }
+}
