+++ massa-pos-exports/src/lib.rs
@@ -9,12 +9,16 @@
 #![feature(let_chains)]
 
 mod controller_traits;
+mod cycle_info;
+mod deferred_credits;
 mod error;
 mod pos_final_state_impl;
 mod settings;
 mod types;
 
 pub use controller_traits::{SelectorController, SelectorManager};
+pub use cycle_info::*;
+pub use deferred_credits::*;
 pub use error::*;
 pub use settings::SelectorConfig;
 pub use types::*;
